+++ app/webpacker/components/ResultsData/Scrambles/ScrambleRowBody.js
@@ -3,8 +3,8 @@ import _ from 'lodash';
 
 import ScrambleRow from './ScrambleRow';
 
-function ScrambleRowBody({ scrambles, adminMode }) {
-  const scramblesByGroupId = Object.values(_.groupBy(scrambles, 'groupId'));
+function ScrambleRowBody({ round, adminMode }) {
+  const scramblesByGroupId = Object.values(_.groupBy(round.scrambles, 'groupId'));
 
   return (
     <>
