--- app/webpacker/stylesheets/competition_scrambles.scss
@@ -1,15 +1,0 @@
-.competition-scrambles {
-  .events-list {
-    display: flex;
-    justify-content: center;
-  }
-  .event-scrambles {
-    .table-cell-solves {
-      word-spacing: 0.5em;
-      vertical-align: middle;
-    }
-  }
-  .prewrap {
-    white-space: pre-wrap;
-  }
-}
